FEAT: Make an interface and make true the interface by qwen(通义千问) Api
