+++ go-service/nexus/handler.go
@@ -6,7 +6,8 @@ import (
 )
 
 // NexusServiceImpl implements the last service interface defined in the IDL.
-type NexusServiceImpl struct{}
+type NexusServiceImpl struct {
+}
 
 func (s *NexusServiceImpl) AskServer(req *nexus_microservice.AskRequest, stream nexus_microservice.NexusService_AskServerServer) (err error) {
 	fmt.Print("EchoServer")
